Done and Undone actions have been implemented for the both types of tasks. Other minor changes were made.

+++ bot/handlers/show_tasks/show_tasks.py
@@ -11,6 +11,7 @@ from handlers.start.start_kb import start_kb
 show_tasks_router = Router()
 db = DataBase()
 
+
 @show_tasks_router.message(
     or_f(
         F.text == '/gettasks',
@@ -18,7 +19,7 @@ db = DataBase()
         )
     )
 async def choose_task_type(message: Message, bot: Bot) -> None:
-   await bot.send_message(message.from_user.id, text='Choose, please:', reply_markup=choose_task_type_kb())
+   await bot.send_message(message.from_user.id, text='Please, make a choice:', reply_markup=choose_task_type_kb())
 
 
 @show_tasks_router.message(F.text == ONGOING_TASKS)
@@ -58,15 +59,31 @@ async def get_completed_tasks_handler(message: Message, bot: Bot) -> None:
         await bot.send_message(message.from_user.id, text=no_tasks_message)
 
 
-@show_tasks_router.callback_query(F.data.startswith('task_delete_'))
-async def delete_task(call: CallbackQuery):
-    user = await db.get_user(call.from_user.id)
-    task_id = call.data.split('_')[-1]
-    try:
-        await db.delete_task(user_id=user.id, task_id=task_id)
-        await call.message.answer(task_deletion_completed)
-    except Exception as error:
-        print(f'Database Deletion Error: {error}')
-        await call.message.answer(error_message)
+@show_tasks_router.callback_query(
+    or_f(
+        F.data.startswith('task_done_'),
+        F.data.startswith('task_undone_'),
+        F.data.startswith('task_delete_')
+    )
+)
+async def task_actions_handler(callback: CallbackQuery) -> None:
+    user = await db.get_user(callback.from_user.id)
+    task_id = callback.data.split('_')[-1]
+    message = None
 
-    
+    try:
+        if (callback.data.startswith('task_done_')):
+            await db.set_task_done(user_id=user.id, task_id=task_id)
+            message = task_setting_done_completed
+        elif (callback.data.startswith('task_undone_')):
+            await db.set_task_undone(user_id=user.id, task_id=task_id)
+            message = task_setting_undone_completed
+        elif (callback.data.startswith('task_delete_')):
+            await db.delete_task(user_id=user.id, task_id=task_id)
+            message = task_deletion_completed
+        
+        await callback.message.answer(message)
+
+    except Exception as error:
+        print(f'Database Query Error: {error}')
+        await callback.message.answer(error_message)
